Removing RuleService and RuleController and merging them to RoomService and RoomController resp
